clang-tidy: a clean tree
The fix-its for the whole check set: override on every overrider (244),
static_cast/reinterpret_cast for the C casts, = default for trivial special
members, nullptr for NULL, make_unique/make_shared for reset(new ...),
range-for where an index only indexed, bool literals. By hand: explicit on
the screens' GuiBase& constructors, PsObj, PsCenterLabel, PsCarouselGame
and OptionsInfo; onScanProgress loses its default arguments (its one caller
passes all four); Opt<T>'s conversion operator is NOLINTed as implicit on
purpose; loop variables named after what they hold; one nested cast the
fix-it mangled (GuiMenuBase::init's row count) rewritten. tools/lint.sh
reports nothing now.

+++ lib_ableem/src/ui/SDL_FontCache.c
@@ -691,12 +691,12 @@ FC_Rect FC_DefaultRenderCallback(FC_Image* src, FC_Rect* srcrect, FC_Target* des
         if(xscale < 0)
         {
             xscale = -xscale;
-            flip = (SDL_RendererFlip) ((int)flip | (int)SDL_FLIP_HORIZONTAL);
+            flip = (SDL_RendererFlip) ((int)flip | SDL_FLIP_HORIZONTAL);
         }
         if(yscale < 0)
         {
             yscale = -yscale;
-            flip = (SDL_RendererFlip) ((int)flip | (int)SDL_FLIP_VERTICAL);
+            flip = (SDL_RendererFlip) ((int)flip | SDL_FLIP_VERTICAL);
         }
 
         SDL_Rect r = *srcrect;
